requirements.txt
streamlit>=1.44.1
st_clickable_images>=0.0.3
streamlit-extras>=0.3.5
openai==1.3.0
httpx==0.24.1
streamlit-image-coordinates>=0.1.6
plotly>=5.17.0
pandas>=2.1.3
streamlit-card>=0.0.61
st-dialog-close-detector>=0.1.0
elevenlabs>=0.2.26
python-dotenv>=1.0.0
fal-client>=0.7.0
requests>=2.31.0
st-paywall>=1.0.1
streamlit-authenticator>=0.2.3
Pillow>=10.1.0
